Replaced the whole buisness logic about user into the view of the user app
