+++ videoflix/videoflix/urls.py
@@ -18,13 +18,15 @@ from django.contrib import admin
 from django.urls import include, path
 from django.conf import settings
 from django.conf.urls.static import static
-from videoflixbackend.views import (
-    LoginView, 
+
+from user.views import (
     SignupView, 
-    LogoutView, 
     VerifyEmailView,
+    LoginView,
+    LogoutView,
     LoggeduserView
 )
+
 
 
 urlpatterns = [
